+++ tests/decorations.test.ts
@@ -1,0 +1,44 @@
+jest.unmock('../src/decorations')
+jest.mock('vscode', () => {
+  return {
+    DecorationRangeBehavior: {
+      ClosedClosed: {},
+    },
+    OverviewRulerLane: {},
+    window: {
+      createTextEditorDecorationType: jest.fn(),
+    },
+  }
+})
+
+import { passingItName, failingItName, skipItName, notRanItName } from '../src/decorations'
+import * as vscode from 'vscode'
+
+describe('Test Result Annotations', () => {
+  describe('Pass', () => {
+    testRangeBehavior(passingItName)
+  })
+
+  describe('Fail', () => {
+    testRangeBehavior(failingItName)
+  })
+
+  describe('Skip', () => {
+    testRangeBehavior(skipItName)
+  })
+
+  describe('Unknown', () => {
+    testRangeBehavior(notRanItName)
+  })
+})
+
+function testRangeBehavior(factoryMethod: Function) {
+  it('should set the range behavior', () => {
+    const mock = vscode.window.createTextEditorDecorationType as jest.Mock<{}>
+    mock.mockReset()
+    factoryMethod()
+
+    expect(mock.mock.calls).toHaveLength(1)
+    expect(mock.mock.calls[0][0].rangeBehavior).toBe(vscode.DecorationRangeBehavior.ClosedClosed)
+  })
+}