+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Master
 
 * Hide Debug CodeLens for skipped tests - seanpoulter
+* Mute output from `console.warn` during tests - seanpoulter
 
 ### 2.5.5