+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Master
 
+* Hide Debug CodeLens for skipped tests - seanpoulter
 * Mute output from `console.warn` during tests - seanpoulter
 
 ### 2.5.5