+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Master
 
+* Improve the dot decoration placement as we edit - seanpoulter
 * Hide Debug CodeLens for skipped tests - seanpoulter
 * Mute output from `console.warn` during tests - seanpoulter